+++ pkg/server/batch_check.go
@@ -102,12 +102,6 @@ func (s *Server) BatchCheck(ctx context.Context, req *openfgav1.BatchCheckReques
 		methodName,
 	).Observe(dispatchCount)
 
-	wasDispatchThrottled := metadata.DispatchThrottleCount > 0
-	if wasDispatchThrottled {
-		throttledRequestCounter.WithLabelValues(s.serviceName, methodName, throttleTypeDispatch).Add(float64(metadata.DispatchThrottleCount))
-	}
-	grpc_ctxtags.Extract(ctx).Set("request.dispatch_throttled", wasDispatchThrottled)
-
 	wasDatastoreThrottled := metadata.DatastoreThrottleCount > 0
 	if wasDatastoreThrottled {
 		throttledRequestCounter.WithLabelValues(s.serviceName, methodName, throttleTypeDatastore).Add(float64(metadata.DatastoreThrottleCount))