+++ pkg/server/commands/check_command_test.go
@@ -227,7 +227,6 @@ type doc
 			DoAndReturn(func(ctx context.Context, req *graph.ResolveCheckRequest) (*graph.ResolveCheckResponse, error) {
 				req.GetRequestMetadata().Depth++
 				req.GetRequestMetadata().DispatchCounter.Add(1)
-				req.GetRequestMetadata().DispatchThrottled.Store(true)
 				ds, _ := storage.RelationshipTupleReaderFromContext(ctx)
 				_, _ = ds.Read(ctx, req.StoreID, storage.ReadFilter{}, storage.ReadOptions{})
 				return nil, context.DeadlineExceeded
@@ -243,7 +242,6 @@ type doc
 		require.Equal(t, uint32(1), checkResp.GetResolutionMetadata().DatastoreQueryCount)
 		require.Equal(t, uint32(1), checkRequestMetadata.Depth)
 		require.Equal(t, uint32(1), checkRequestMetadata.DispatchCounter.Load())
-		require.True(t, checkRequestMetadata.DispatchThrottled.Load())
 	})
 }