+++ pkg/server/server_test.go
@@ -43,7 +43,6 @@ import (
 	"github.com/openfga/openfga/pkg/storage/sqlcommon"
 	"github.com/openfga/openfga/pkg/storage/sqlite"
 	"github.com/openfga/openfga/pkg/storage/storagewrappers"
-	storageTest "github.com/openfga/openfga/pkg/storage/test"
 	storagefixtures "github.com/openfga/openfga/pkg/testfixtures/storage"
 	"github.com/openfga/openfga/pkg/testutils"
 	"github.com/openfga/openfga/pkg/tuple"
@@ -175,20 +174,6 @@ func TestServerPanicIfValidationsFail(t *testing.T) {
 		})
 	})
 
-	t.Run("invalid_dispatch_throttle_threshold", func(t *testing.T) {
-		require.PanicsWithError(t, "failed to construct the OpenFGA server: check default dispatch throttling threshold must be equal or smaller than max dispatch threshold for Check", func() {
-			mockController := gomock.NewController(t)
-			defer mockController.Finish()
-			mockDatastore := mockstorage.NewMockOpenFGADatastore(mockController)
-			_ = MustNewServerWithOpts(
-				WithDatastore(mockDatastore),
-				WithDispatchThrottlingCheckResolverEnabled(true),
-				WithDispatchThrottlingCheckResolverThreshold(100),
-				WithDispatchThrottlingCheckResolverMaxThreshold(80),
-			)
-		})
-	})
-
 	t.Run("invalid_access_control_setup", func(t *testing.T) {
 		require.PanicsWithError(t, "failed to construct the OpenFGA server: access control parameters are not enabled. They can be enabled for experimental use by passing the `--experimentals enable-access-control` configuration option when running OpenFGA server. Additionally, the `--access-control-store-id` and `--access-control-model-id` parameters must not be empty", func() {
 			mockController := gomock.NewController(t)
@@ -309,47 +294,6 @@ func TestServerPanicIfEmptyRequestDurationDispatchCountBuckets(t *testing.T) {
 	})
 }
 
-func TestServerPanicIfDefaultDispatchThresholdGreaterThanMaxDispatchThreshold(t *testing.T) {
-	require.PanicsWithError(t, "failed to construct the OpenFGA server: check default dispatch throttling threshold must be equal or smaller than max dispatch threshold for Check", func() {
-		mockController := gomock.NewController(t)
-		defer mockController.Finish()
-		mockDatastore := mockstorage.NewMockOpenFGADatastore(mockController)
-		_ = MustNewServerWithOpts(
-			WithDatastore(mockDatastore),
-			WithDispatchThrottlingCheckResolverEnabled(true),
-			WithDispatchThrottlingCheckResolverThreshold(100),
-			WithDispatchThrottlingCheckResolverMaxThreshold(80),
-		)
-	})
-}
-
-func TestServerPanicIfDefaultListObjectsThresholdGreaterThanMaxDispatchThreshold(t *testing.T) {
-	require.PanicsWithError(t, "failed to construct the OpenFGA server: ListObjects default dispatch throttling threshold must be equal or smaller than max dispatch threshold for ListObjects", func() {
-		mockController := gomock.NewController(t)
-		defer mockController.Finish()
-		mockDatastore := mockstorage.NewMockOpenFGADatastore(mockController)
-		_ = MustNewServerWithOpts(
-			WithDatastore(mockDatastore),
-			WithListObjectsDispatchThrottlingEnabled(true),
-			WithListObjectsDispatchThrottlingThreshold(100),
-			WithListObjectsDispatchThrottlingMaxThreshold(80),
-		)
-	})
-}
-
-func TestServerPanicIfDefaultListUsersThresholdGreaterThanMaxDispatchThreshold(t *testing.T) {
-	require.PanicsWithError(t, "failed to construct the OpenFGA server: ListUsers default dispatch throttling threshold must be equal or smaller than max dispatch threshold for ListUsers", func() {
-		mockController := gomock.NewController(t)
-		defer mockController.Finish()
-		mockDatastore := mockstorage.NewMockOpenFGADatastore(mockController)
-		_ = MustNewServerWithOpts(
-			WithDatastore(mockDatastore),
-			WithListUsersDispatchThrottlingEnabled(true),
-			WithListUsersDispatchThrottlingThreshold(100),
-			WithListUsersDispatchThrottlingMaxThreshold(80),
-		)
-	})
-}
 func TestServerWithPostgresDatastore(t *testing.T) {
 	t.Cleanup(func() {
 		goleak.VerifyNone(t)
@@ -1705,9 +1649,6 @@ func TestDelegateCheckResolver(t *testing.T) {
 	})
 	t.Run("default_check_resolver_alone", func(t *testing.T) {
 		cfg := serverconfig.DefaultConfig()
-		require.False(t, cfg.CheckDispatchThrottling.Enabled)
-		require.False(t, cfg.ListObjectsDispatchThrottling.Enabled)
-		require.False(t, cfg.ListUsersDispatchThrottling.Enabled)
 		require.False(t, cfg.CheckQueryCache.Enabled)
 
 		ds := memory.New()
@@ -1716,7 +1657,6 @@ func TestDelegateCheckResolver(t *testing.T) {
 			WithDatastore(ds),
 		)
 		t.Cleanup(s.Close)
-		require.False(t, s.checkDispatchThrottlingEnabled)
 
 		require.False(t, s.cacheSettings.CheckQueryCacheEnabled)
 
@@ -1731,99 +1671,6 @@ func TestDelegateCheckResolver(t *testing.T) {
 		require.True(t, ok)
 	})
 
-	t.Run("dispatch_throttling_check_resolver_enabled", func(t *testing.T) {
-		ds := memory.New()
-		t.Cleanup(ds.Close)
-		const dispatchThreshold = 50
-		s := MustNewServerWithOpts(
-			WithDatastore(ds),
-			WithDispatchThrottlingCheckResolverEnabled(true),
-			WithDispatchThrottlingCheckResolverThreshold(dispatchThreshold),
-		)
-		t.Cleanup(s.Close)
-
-		require.False(t, s.cacheSettings.CheckQueryCacheEnabled)
-
-		require.True(t, s.checkDispatchThrottlingEnabled)
-		require.EqualValues(t, dispatchThreshold, s.checkDispatchThrottlingDefaultThreshold)
-		require.EqualValues(t, 0, s.checkDispatchThrottlingMaxThreshold)
-		checkResolver, closer, _ := s.getCheckResolverBuilder("store_id_123").Build()
-		defer closer()
-		require.NotNil(t, checkResolver)
-
-		dispatchThrottlingResolver, ok := checkResolver.(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-
-		localChecker, ok := dispatchThrottlingResolver.GetDelegate().(*graph.LocalChecker)
-		require.True(t, ok)
-
-		_, ok = localChecker.GetDelegate().(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-	})
-
-	t.Run("dispatch_throttling_check_resolver_enabled_zero_max_threshold", func(t *testing.T) {
-		ds := memory.New()
-		t.Cleanup(ds.Close)
-		const dispatchThreshold = 50
-		s := MustNewServerWithOpts(
-			WithDatastore(ds),
-			WithDispatchThrottlingCheckResolverEnabled(true),
-			WithDispatchThrottlingCheckResolverThreshold(dispatchThreshold),
-			WithDispatchThrottlingCheckResolverMaxThreshold(0),
-		)
-		t.Cleanup(s.Close)
-
-		require.False(t, s.cacheSettings.CheckQueryCacheEnabled)
-
-		require.True(t, s.checkDispatchThrottlingEnabled)
-		require.EqualValues(t, dispatchThreshold, s.checkDispatchThrottlingDefaultThreshold)
-		require.EqualValues(t, 0, s.checkDispatchThrottlingMaxThreshold)
-		checkResolver, closer, _ := s.getCheckResolverBuilder("store_id_123").Build()
-		defer closer()
-		require.NotNil(t, checkResolver)
-
-		dispatchThrottlingResolver, ok := checkResolver.(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-
-		localChecker, ok := dispatchThrottlingResolver.GetDelegate().(*graph.LocalChecker)
-		require.True(t, ok)
-
-		_, ok = localChecker.GetDelegate().(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-	})
-
-	t.Run("dispatch_throttling_check_resolver_enabled_non_zero_max_threshold", func(t *testing.T) {
-		ds := memory.New()
-		t.Cleanup(ds.Close)
-		const dispatchThreshold = 50
-		const maxDispatchThreshold = 60
-
-		s := MustNewServerWithOpts(
-			WithDatastore(ds),
-			WithDispatchThrottlingCheckResolverEnabled(true),
-			WithDispatchThrottlingCheckResolverThreshold(dispatchThreshold),
-			WithDispatchThrottlingCheckResolverMaxThreshold(maxDispatchThreshold),
-		)
-		t.Cleanup(s.Close)
-
-		require.False(t, s.cacheSettings.CheckQueryCacheEnabled)
-
-		require.True(t, s.checkDispatchThrottlingEnabled)
-		require.EqualValues(t, dispatchThreshold, s.checkDispatchThrottlingDefaultThreshold)
-		require.EqualValues(t, maxDispatchThreshold, s.checkDispatchThrottlingMaxThreshold)
-		checkResolver, closer, _ := s.getCheckResolverBuilder("store_id_123").Build()
-		defer closer()
-		require.NotNil(t, checkResolver)
-		dispatchThrottlingResolver, ok := checkResolver.(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-
-		localChecker, ok := dispatchThrottlingResolver.GetDelegate().(*graph.LocalChecker)
-		require.True(t, ok)
-
-		_, ok = localChecker.GetDelegate().(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-	})
-
 	t.Run("cache_check_resolver_enabled", func(t *testing.T) {
 		ds := memory.New()
 		t.Cleanup(ds.Close)
@@ -1832,8 +1679,6 @@ func TestDelegateCheckResolver(t *testing.T) {
 			WithCheckQueryCacheEnabled(true),
 		)
 		t.Cleanup(s.Close)
-
-		require.False(t, s.checkDispatchThrottlingEnabled)
 
 		require.True(t, s.cacheSettings.CheckQueryCacheEnabled)
 		checkResolver, closer, _ := s.getCheckResolverBuilder("store_id_123").Build()
@@ -1844,40 +1689,6 @@ func TestDelegateCheckResolver(t *testing.T) {
 		require.True(t, ok)
 
 		localChecker, ok := cachedCheckResolver.GetDelegate().(*graph.LocalChecker)
-		require.True(t, ok)
-
-		_, ok = localChecker.GetDelegate().(*graph.CachedCheckResolver)
-		require.True(t, ok)
-	})
-
-	t.Run("both_dispatch_throttling_and_cache_check_resolver_enabled", func(t *testing.T) {
-		ds := memory.New()
-		t.Cleanup(ds.Close)
-		s := MustNewServerWithOpts(
-			WithDatastore(ds),
-			WithCheckQueryCacheEnabled(true),
-			WithDispatchThrottlingCheckResolverEnabled(true),
-			WithDispatchThrottlingCheckResolverThreshold(50),
-			WithDispatchThrottlingCheckResolverMaxThreshold(100),
-		)
-		t.Cleanup(s.Close)
-
-		require.True(t, s.checkDispatchThrottlingEnabled)
-		require.EqualValues(t, 50, s.checkDispatchThrottlingDefaultThreshold)
-		require.EqualValues(t, 100, s.checkDispatchThrottlingMaxThreshold)
-
-		checkResolver, closer, _ := s.getCheckResolverBuilder("store_id_123").Build()
-		defer closer()
-		require.NotNil(t, checkResolver)
-
-		cachedCheckResolver, ok := checkResolver.(*graph.CachedCheckResolver)
-		require.True(t, ok)
-		require.True(t, s.cacheSettings.CheckQueryCacheEnabled)
-
-		dispatchThrottlingResolver, ok := cachedCheckResolver.GetDelegate().(*graph.DispatchThrottlingCheckResolver)
-		require.True(t, ok)
-
-		localChecker, ok := dispatchThrottlingResolver.GetDelegate().(*graph.LocalChecker)
 		require.True(t, ok)
 
 		_, ok = localChecker.GetDelegate().(*graph.CachedCheckResolver)
@@ -1949,96 +1760,6 @@ func TestIsAccessControlEnabled(t *testing.T) {
 		)
 		t.Cleanup(s.Close)
 		require.True(t, s.IsAccessControlEnabled())
-	})
-}
-
-func TestServer_ThrottleUntilDeadline(t *testing.T) {
-	t.Cleanup(func() {
-		goleak.VerifyNone(t)
-	})
-	ds := memory.New()
-	t.Cleanup(ds.Close)
-
-	modelStr := `
-		model
-			schema 1.1
-		type user
-
-		type group
-		relations
-			define other: [user]
-			define member: [user, group#member, group#other]
-
-		type document
-		relations
-			define viewer: [user, group#member]`
-
-	tuples := []string{
-		"document:1#viewer@user:jon", // Observed before first dispatch
-		"document:1#viewer@group:eng#member",
-		"group:eng#member@group:backend#member",
-		"group:backend#member@user:tyler", // Requires two dispatches, gets throttled
-
-		"document:2#viewer@user:tyler",
-	}
-
-	storeID, model := storageTest.BootstrapFGAStore(t, ds, modelStr, tuples)
-	t.Cleanup(ds.Close)
-
-	deadline := 50 * time.Millisecond
-
-	s := MustNewServerWithOpts(
-		WithDatastore(ds),
-
-		WithDispatchThrottlingCheckResolverEnabled(true),
-		WithDispatchThrottlingCheckResolverFrequency(3*deadline), // Forces time-out when throttling occurs
-		WithDispatchThrottlingCheckResolverThreshold(1),          // Applies throttling after first dispatch
-
-		WithListObjectsDeadline(deadline),
-		WithListObjectsDispatchThrottlingEnabled(true),
-		WithListObjectsDispatchThrottlingThreshold(1),          // Applies throttling after first dispatch
-		WithListObjectsDispatchThrottlingFrequency(3*deadline), // Forces time-out when throttling occurs
-
-		WithListUsersDeadline(deadline),
-		WithListUsersDispatchThrottlingEnabled(true),
-		WithListUsersDispatchThrottlingThreshold(1),          // Applies throttling after first dispatch
-		WithListUsersDispatchThrottlingFrequency(2*deadline), // Forces time-out when throttling occurs
-	)
-	t.Cleanup(s.Close)
-
-	ctx := context.Background()
-
-	t.Run("list_users_return_no_error_and_partial_results", func(t *testing.T) {
-		resp, err := s.ListUsers(ctx, &openfgav1.ListUsersRequest{
-			StoreId:              storeID,
-			AuthorizationModelId: model.GetId(),
-			Object: &openfgav1.Object{
-				Type: "document",
-				Id:   "1",
-			},
-			Relation: "viewer",
-			UserFilters: []*openfgav1.UserTypeFilter{
-				{Type: "user"},
-			},
-		})
-
-		require.NoError(t, err)
-		require.NotNil(t, resp)
-		require.LessOrEqual(t, len(resp.GetUsers()), 1) // race condition of context cancellation
-	})
-
-	t.Run("list_objects_return_no_error_and_partial_results", func(t *testing.T) {
-		resp, err := s.ListObjects(ctx, &openfgav1.ListObjectsRequest{
-			StoreId:              storeID,
-			AuthorizationModelId: model.GetId(),
-			User:                 "user:tyler",
-			Relation:             "viewer",
-			Type:                 "document",
-		})
-
-		require.NoError(t, err)
-		require.NotNil(t, resp)
-		require.LessOrEqual(t, len(resp.GetObjects()), 1) // race condition of context cancellation
 	})
 }